fix(lens): relocate notifyDirty to publisher.ts beside its channel
Move the guarded notifyDirty emitter from autosaveGate.ts (one of the pulse's two consumers) into publisher.ts next to terminalsDirtyChannel, point the three producers (terminals.ts, terminalEndpoint/metadata.ts, terminalEndpoint/local.ts) and the assembly barrel at the neutral channel registry, and leave autosaveGate.ts a pure subscriber.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-2, raised by lowy). Not pushed or merged.
